refactor(interpreter): extract word/parameter expansion into submodule
Move the expand_word/expand_word_to_fields pipeline, parameter-expansion
operators (apply_parameter_op, replace_pattern, remove_pattern*,
apply_param_op_maybe_per_element), IFS field splitting, operand quoting,
and pattern helpers out of interpreter/mod.rs into interpreter/expansion.rs
(~1700 lines). Command-substitution/subshell-snapshot machinery stays in the
parent. MAX_EXPANSION_RESULT_BYTES promoted to pub(crate) for sibling access.
